backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-decouple
requests
psycopg2-binary
